+++ src/os/os.c
@@ -12,5 +12,5 @@ void os_terminate() {
 static i8 shared_buffer[512];
 ustring os_cwd() {
     getcwd(shared_buffer, 512);
-    return (ustring){.data = shared_buffer, .length = strlen(shared_buffer)};
+    return (ustring){ .data = shared_buffer, .length = (u32)strlen(shared_buffer) };
 }
